requirements.txt
pycryptodome
clipboard==0.0.4
flask==1.1.2
image==1.5.33
qrcode==6.1
